Obavljen 6.zadatak i bolje organiziran kod
