+++ app/src/main/java/agency/five/codebase/android/movieapp/ui/component/FavoriteButton.kt
@@ -4,6 +4,7 @@ package agency.five.codebase.android.movieapp.ui.component
 import agency.five.codebase.android.movieapp.R
 import agency.five.codebase.android.movieapp.ui.theme.Blue
 import agency.five.codebase.android.movieapp.ui.theme.Favorite
+import agency.five.codebase.android.movieapp.ui.theme.MovieAppTheme
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -15,6 +16,7 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
 @Composable
@@ -41,6 +43,14 @@ fun FavoriteButton(
         )
     }
 
+}
+
+@Preview
+@Composable
+private fun FavoriteButtonPreview(){
+    MovieAppTheme {
+        FavoriteButton()
+    }
 }
