+++ app/src/main/java/agency/five/codebase/android/movieapp/ui/component/UserScoreProgressBar.kt
@@ -1,21 +1,25 @@
 package agency.five.codebase.android.movieapp.ui.component
 
+import agency.five.codebase.android.movieapp.ui.theme.MovieAppTheme
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+
+
 
 @Composable
 fun UserScoreProgressBar(
@@ -39,7 +43,7 @@ fun UserScoreProgressBar(
     }
     Box(
         contentAlignment = Alignment.Center,
-        modifier = Modifier.size(radius * 2) // diameter
+        modifier = Modifier.size(radius * 2)
     ) {
 
         Canvas(modifier = Modifier.size(radius)) {
@@ -54,7 +58,21 @@ fun UserScoreProgressBar(
                 ),
             )
         }
-
-        Text(text = "${percentage*10}")
+        Surface(
+            color = MaterialTheme.colors.background,
+        ) {
+            Text(text = "${percentage * 10}")
+        }
     }
 }
+
+
+
+@Preview
+@Composable
+private fun UserScorePRogressBarPreview(){
+    MovieAppTheme {
+        UserScoreProgressBar(percentage = 0.68f)
+    }
+}
+
